use geojson endpoint, reverse coordinate order
One of the changes includes using a geojson specific endpoint which, depending on if the geometry is polygon or multipolygon, could have coordinates of a 3 or 4 dimensional array, respectively. I used https://github.com/paulmach/go.geojson to take care of unmarshaling a json body into a featurecollection while using a 3 or 4 dimensional array appropriately

The arcgis geojson endpoint isn't adhering to RFC 7946 standards regarding the right-hand wrap protocol. Specifically, for a given polygon, the first/exerior ring must wrap in a counterclockwise direction while all rings following the first/exterior ring must wrap in a clockwise direction.

Arcgis data was doing the opposite, so I created two data slices containing arcgis data, and reassigned the values for one of them to reverse the direction that the coordinates were layed out in a ring. The order of the rings themselves within a polygon should remain the same, since arcgis is correctly putting the exterior ring first with all interior rings following the first ring.